requirements.txt
cassio
langchain
langchain-community
langchain_groq
langchain_huggingface
langchain_core
langchain_tavily
python-dotenv
pypdf
faiss-cpu
ipykernel
chromadb
langgraph
langgraph-checkpoint-sqlite
sentence-transformers
fastapi
pydantic[email]
flask